+++ src/main/java/org/stellar/sdk/responses/operations/ManageSellOfferOperationResponse.java
@@ -13,7 +13,7 @@ import org.stellar.sdk.AssetTypeNative;
  */
 public class ManageSellOfferOperationResponse extends OperationResponse {
   @SerializedName("offer_id")
-  protected final Integer offerId;
+  protected final Long offerId;
   @SerializedName("amount")
   protected final String amount;
   // Price is not implemented yet in horizon
@@ -35,7 +35,7 @@ public class ManageSellOfferOperationResponse extends OperationResponse {
   @SerializedName("selling_asset_issuer")
   protected final String sellingAssetIssuer;
 
-  ManageSellOfferOperationResponse(Integer offerId, String amount, String price, String buyingAssetType, String buyingAssetCode, String buyingAssetIssuer, String sellingAssetType, String sellingAssetCode, String sellingAssetIssuer) {
+  ManageSellOfferOperationResponse(Long offerId, String amount, String price, String buyingAssetType, String buyingAssetCode, String buyingAssetIssuer, String sellingAssetType, String sellingAssetCode, String sellingAssetIssuer) {
     this.offerId = offerId;
     this.amount = amount;
     this.price = price;
@@ -47,7 +47,7 @@ public class ManageSellOfferOperationResponse extends OperationResponse {
     this.sellingAssetIssuer = sellingAssetIssuer;
   }
 
-  public Integer getOfferId() {
+  public Long getOfferId() {
     return offerId;
   }
